Fix Google

+++ FALOFinancialProofing.FALOHomePage/Models/CampaignDTO.cs
@@ -8,8 +8,8 @@
         public string Title { get; set; }
         public string Description { get; set; }
         public DateTime DateOfCreation { get; set; }
-        public decimal FundTarget { get; set; }
-        public decimal? TotalMoneyEarned { get; set; }
+        public long FundTarget { get; set; }
+        public long? TotalMoneyEarned { get; set; }
         public string? Image { get; set; }
         public DateTime? EndDate { get; set; }
         public string Address { get; set; }
